feat(restructure-controller): Refactored change password route
